+++ app/Http/Controllers/Api/CategoryController.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace App\Http\Controllers\Api;
+
+use App\Http\Controllers\Controller;
+use App\Models\Category;
+use Illuminate\Http\Request;
+
+class CategoryController extends Controller
+{
+    private $rules = [
+        'name' => 'required|max:255',
+        'is_active' => 'boolean'
+    ];
+    public function index()
+    {
+        return Category::all();
+    }
+
+    public function store(Request $request)
+    {
+        $this->validate($request, $this->rules);
+        Category::craete($request->all());
+    }
+
+    public function show(Category $category)
+    {
+        return $category;
+    }
+
+    public function update(Request $request, Category $category)
+    {
+        $this->validate($request, $this->rules);
+        $category->update($request->all());
+    }
+
+    public function destroy(Category $category)
+    {
+        $category->delete();
+        return response()->noContent();//204 - NO content
+    }
+}
