+++ app/Models/Category.php
@@ -6,5 +6,9 @@ use Illuminate\Database\Eloquent\Model;
 
 class Category extends Model
 {
-    //
+    protected $fillable = [
+        'name',
+        'description',
+        'is_active'
+    ];
 }
